ref: separate snowflake tests to letsql/backends/snowflake/tests
