+++ python/letsql/backends/let/tests/test_cache.py
@@ -1,32 +1,26 @@
 from __future__ import annotations
 
 import pathlib
-import re
 import time
 
 import ibis
-import pandas as pd
 import pytest
 from ibis import _
-from ibis.util import gen_name
 
 import letsql
-import letsql as ls
 from letsql.backends.let import (
     Backend,
 )
-from letsql.backends.let.tests.conftest import (
-    assert_frame_equal,
+from letsql.backends.conftest import (
     get_storage_uncached,
-    inside_temp_schema,
 )
 from letsql.common.caching import KEY_PREFIX
 from letsql.common.utils.postgres_utils import (
     do_analyze,
     get_postgres_n_scans,
 )
-from letsql.common.utils.snowflake_utils import (
-    get_session_query_df,
+from letsql.tests.util import (
+    assert_frame_equal,
 )
 
 
@@ -351,49 +345,3 @@ def test_postgres_cache_invalidation(pg, con):
     modify_postgres_table(dt)
     expr_cached.execute()
     assert_n_scans_changes(dt, n_scans_after)
-
-
-@pytest.mark.snowflake
-def test_snowflake_cache_invalidation(con, sf_con, temp_catalog, temp_db, tmp_path):
-    group_by = "key"
-    df = pd.DataFrame({group_by: list("abc"), "value": [1, 2, 3]})
-    name = gen_name("tmp_table")
-    storage = ls.common.caching.ParquetCacheStorage(tmp_path, source=con)
-
-    # must explicitly invoke USE SCHEMA: use of temp_* DOESN'T impact internal CREATE TEMP STAGE
-    with inside_temp_schema(sf_con, temp_catalog, temp_db):
-        table = sf_con.create_table(
-            name=name,
-            obj=df,
-        )
-        t = con.register(table, f"let_{table.op().name}")
-        cached_expr = (
-            t.group_by(group_by)
-            .agg({f"min_{col}": t[col].min() for col in t.columns})
-            .cache(storage)
-        )
-        (storage, uncached) = get_storage_uncached(con, cached_expr)
-        unbound_sql = re.sub(
-            r"\s+",
-            " ",
-            ibis.to_sql(uncached, dialect=sf_con.name),
-        )
-        query_df = get_session_query_df(sf_con)
-
-        # test preconditions
-        assert not storage.exists(uncached)
-        assert query_df.QUERY_TEXT.eq(unbound_sql).sum() == 0
-
-        # test cache creation
-        cached_expr.execute()
-        query_df = get_session_query_df(sf_con)
-        assert storage.exists(uncached)
-        assert query_df.QUERY_TEXT.eq(unbound_sql).sum() == 1
-
-        # test cache use
-        cached_expr.execute()
-        assert query_df.QUERY_TEXT.eq(unbound_sql).sum() == 1
-
-        # test cache invalidation
-        sf_con.insert(name, df, database=f"{temp_catalog}.{temp_db}")
-        assert not storage.exists(uncached)
