+++ python/letsql/backends/let/tests/test_client.py
@@ -1,7 +1,9 @@
 import pyarrow as pa
 
 import letsql as ls
-from letsql.backends.let.tests.conftest import assert_frame_equal
+from letsql.tests.util import (
+    assert_frame_equal,
+)
 
 
 def test_register_record_batch_reader(alltypes_df):
